% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseLocalAlignment-package.R
\docType{package}
\name{pairwiseLocalAlignment-package}
\alias{pairwiseLocalAlignment-package}
\alias{pairwiseLocalAlignment}
\title{Pairwise Local Alignment of nucleotide sequences}
\description{
The \code{pairwiseLocalAlignment} package implements a function to perform
pairwise local alignment of nucleotide sequences implementing the
Smith-Waterman algorithm. For further information check the \code{smith_waterman}
documentation with \code{?smith_waterman}
}
\details{
\tabular{ll}{
Package: \tab pairwiseLocalAlignment\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2023-01-10\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Alessandro Gennari\cr
Politecnico di Milano\cr
Maintainer: Alessandro Gennari\cr
E-Mail: \href{mailto:alessandro.gennari@mail.polimi.it}{alessandro.gennari@mail.polimi.it}
}
