% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smith_waterman}
\alias{smith_waterman}
\title{Pairwise Local Alignment}
\usage{
smith_waterman(match_cost, mismatch_cost, gap_cost, seq1, seq2)
}
\arguments{
\item{match_cost}{score of a match between two nucleotides}

\item{mismatch_cost}{score of a mismatch between two nucleotides}

\item{gap_cost}{score of the insertion gap between two nucleotides}

\item{seq1}{first nucletide sequence}

\item{seq2}{second nucletide sequence}
}
\value{
list with two elements: s1 and s2 which correspond to the two
sub-sequences resulting from the pairwise local alignment
}
\description{
This function reads 3 numerical parameters which defines the cost of a
match, mismatch or the insertion of a gap between nucleotides
of two sequences aligned. Then it perform the Smith-Waterman algorithm for
Pairwise Local Alignment of sequences and return an list collecting the two
aligned sub-sequences.
}
\examples{

smith_waterman(10, -5, -6, "GAATC", "CATACG")

}
\references{
\url{https://en.wikipedia.org/wiki/Smith–Waterman_algorithm}\cr
}
\author{
Alessandro Gennari\cr Politecnico di Milano\cr Maintainer: Alessandro
Gennari\cr E-Mail: \href{mailto:alessandro.gennari@mail.polimi.it}{alessandro.gennari@mail.polimi.it}
}
